swapped code between Terminal.vue and TerminalEmulator.vue Added full screen feature on both terminals
